refactor(auth): canonicalize totp token and preference modules
