+++ packages/opencode/src/auth/totp-preference.ts
@@ -1,0 +1,1 @@
+export * from "@opencode-ai/fork-auth/auth/totp-preference"
