+++ packages/opencode/src/auth/two-factor-preference.ts
@@ -1,1 +1,1 @@
-export * from "@opencode-ai/fork-auth/auth/two-factor-preference"
+export * from "./totp-preference"
